requirements.txt
aiocron==1.8
catboost==1.2.8
fastapi==0.115.7
httpx==0.28.1
pandas==2.2.3
plotly==5.24.1
pymysql==1.1.0
cryptography==41.0.7
pydantic==2.10.5
python-dotenv==1.0.1
requests==2.32.3
scikit-learn==1.6.0
SQLAlchemy==2.0.37
streamlit==1.41.1
uvicorn==0.34.0
lightgbm==4.1.0
mlflow==2.8.1
dvc==3.66.0
dvc-s3==3.1.0
pytest==7.4.3
black==23.12.1
flake8==7.0.0
selenium==4.15.2
webdriver-manager==4.0.1
playwright==1.40.0
nest_asyncio==1.6.0
awscli==1.32.0


